feat: 주문 완료하기 API 구현 (#472)
* refactor: Money의 정적 팩토리 생성자가 Long을 지원하도록 개선

* feat: 주문 완료 로직 구현

* feat: 주문 완료 이벤트 발행 로직 추가

* feat: 주문 완료 여부 조회 로직 추가

* docs: 투두 추가

* feat: orderId 대신 nanoId 사용하도록 변경

* feat: 주문 완료 검증 로직 추가

* style: 개행 제거

* feat: nanoId 주문 조회 로직 추가

* feat: 주문 완료 서비스 구현

* refactor: nested 적용

* docs: 로그 추가

* test: 주문 완료 검증 테스트 추가

* test: 주문 완료 통합 테스트 추가

* docs: 투두 추가

* refactor: given 절로 이동

+++ src/test/java/com/gdschongik/gdsc/domain/common/vo/MoneyTest.java
@@ -14,8 +14,8 @@ class MoneyTest {
         @Test
         void 값과_스케일_모두_같으면_동일한_금액이다() {
             // given
-            Money money1 = Money.from(BigDecimal.valueOf(1000));
-            Money money2 = Money.from(BigDecimal.valueOf(1000));
+            Money money1 = Money.from(1000L);
+            Money money2 = Money.from(1000L);
 
             // when & then
             assertThat(money1).isEqualTo(money2);
@@ -24,7 +24,7 @@ class MoneyTest {
         @Test
         void 스케일이_달라도_같은_값이면_동일한_금액이다() {
             // given
-            Money money1 = Money.from(BigDecimal.valueOf(1000));
+            Money money1 = Money.from(1000L);
             Money money2 = Money.from(BigDecimal.valueOf(1000.0));
             Money money3 = Money.from(BigDecimal.valueOf(1000.00));
             Money money4 = Money.from(BigDecimal.valueOf(1000.000));
@@ -56,8 +56,8 @@ class MoneyTest {
         @Test
         void 값과_스케일_모두_같으면_동일한_해시코드이다() {
             // given
-            Money money1 = Money.from(BigDecimal.valueOf(1000));
-            Money money2 = Money.from(BigDecimal.valueOf(1000));
+            Money money1 = Money.from(1000L);
+            Money money2 = Money.from(1000L);
 
             // when & then
             int expected = money2.hashCode();
@@ -67,7 +67,7 @@ class MoneyTest {
         @Test
         void 스케일이_달라도_같은_값이면_동일한_해시코드이다() {
             // given
-            Money money1 = Money.from(BigDecimal.valueOf(1000));
+            Money money1 = Money.from(1000L);
             Money money2 = Money.from(BigDecimal.valueOf(1000.0));
             Money money3 = Money.from(BigDecimal.valueOf(1000.00));
             Money money4 = Money.from(BigDecimal.valueOf(1000.000));
